refactor: extract board share module
Resolves the last remaining temporary circular import: alerts.mjs now
imports getShareState/toggleShare from lib/share.mjs directly instead of
../server.mjs. No circular imports or TEMPORARY export markers remain
anywhere in packages/board.
